minor PR cleanups + refactor out common test util classes
