+++ services/s3/src/main/java/software/amazon/awssdk/services/s3/internal/multipart/UnknownContentLengthAsyncRequestBodySubscriber.java
@@ -271,14 +271,7 @@ public class UnknownContentLengthAsyncRequestBodySubscriber implements Subscribe
     }
 
     private void completeMultipartUploadIfFinish() {
-        // asyncRequestBodyInFlight MUST be re-read here rather than passed in by the caller. In the upload
-        // completion callback, subscription.request(1) is invoked between decrementAndGet() and this method,
-        // which can synchronously deliver the final AsyncRequestBody (starting a new upload) followed by
-        // onComplete() (setting isDone). A stale snapshot of 0 taken before those deliveries would then
-        // initiate CompleteMultipartUpload while the final part is still in flight, completing the upload
-        // with a missing part. Reading isDone (volatile) before the counter guarantees that, once isDone is
-        // observed as true, all onNext() increments are visible, so a fresh read of 0 means every upload
-        // that was started has finished and its CompletedPart is visible in completedParts.
+        // All atomics including asyncRequestBodyInFlight MUST be re-read here rather than passed in by the caller.
         if (isDone && asyncRequestBodyInFlight.get() == 0 && completedMultipartInitiated.compareAndSet(false, true)) {
             CompletedPart[] parts = completedParts.stream()
                                                   .sorted(Comparator.comparingInt(CompletedPart::partNumber))
